feat: Modify layout and routes to support AI assistant
We leverage outlet like on other apps to make
possible to display AI assistant *above* the iframe
so when we open and close AI assistant it does not
reload the iframe.

We also use a prefix in iframe routing leveraging
cozy-external-bridge 0.9.0.

We also migrate to HashRouter to stay consistent
with cozy-twakemail.